Comentarios hechos del App.jsx y cambiado extension
